SnapdragonCamera: Remove Camera2 apk.
Override Camera2 with SnapdragonCamera to remove Camera2 from
build. This is to avoid confusion to user with similar camera
apps in the build. Finally, there will be only two camera
applications - SnapdragonCamera, which is the QC value added
Camera app and another GoogleCamera, which is the stock Camera
app provided by Google.

Change-Id: Ifc50772848b5a5e9a64bc0371bd6980d8633e8ac