Reset preview format to nv21 for non 1080p/720p resolutions
Camera application is setting nv12_venus as preview format when user
choose 1080p or 720p as video resolution. And preview format is not
changed when user switches to other resolutions. Due to which we see
preview flicker for smaller video resolutions in few targets. This
change will reset preview format to yuv420sp to avoid preview flicker.

Change-Id: I92d644c8369115d69a5ecc95ba31d8f2fb52ae0a