Update patch set 1
Patch Set 1:

But thats why I named it "external storage" because you cant even find out if its a SD card or USB storage. And more then one is also not supported. We can hide the menu if you really want but then need to test what happens if you add storage "on the fly" - in worst case app needs to be restarted.

Patch-set: 1