Update patch set 8
Patch Set 8:

> Humberto, So reverting that commit would give us the correct
 > orientation back? Did you try tat already?

Nope. I mean it will break portrait orientation. That patch means that Snapdragoncamera works better on portrait devices for now. We need something like that implementation for landscape devices.

As I told, this patch is locking orientation as landscape and it is a wrong way to get it fixed.

Patch-set: 8